test: complete ut for magic env var
