+++ pkg/generator/appconfiguration/util.go
@@ -2,9 +2,7 @@ package appconfiguration
 
 import (
 	"sort"
-	"strings"
 
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"kusionstack.io/kusion/pkg/models"
@@ -126,131 +124,5 @@ func UniqueAppLabels(projectName, appName string) map[string]string {
 	return map[string]string{
 		"app.kubernetes.io/part-of": projectName,
 		"app.kubernetes.io/name":    appName,
-	}
-}
-
-// MagicEnvVar generates a specialized EnvVar based on the key and
-// value of environment.
-func MagicEnvVar(k, v string) *corev1.EnvVar {
-	supportedParsers := []MagicEnvParser{
-		SecretEnvParser,
-		ConfigMapEnvParser,
-		RawEnvParser,
-	}
-	for _, p := range supportedParsers {
-		if p.Match(k, v) {
-			return p.Gen(k, v)
-		}
-	}
-	return nil
-}
-
-var (
-	SecretEnvParser    MagicEnvParser = NewSecretEnvParser()
-	ConfigMapEnvParser                = NewConfigMapEnvParser()
-	RawEnvParser                      = NewRawEnvParser()
-)
-
-// MagicEnvParser is an interface for environment variable parsers.
-type MagicEnvParser interface {
-	Match(k, v string) (matched bool)
-	Gen(k, v string) *corev1.EnvVar
-}
-
-// rawEnvParser is a parser for raw environment variables.
-type rawEnvParser struct{}
-
-// NewRawEnvParser creates a new instance of RawEnvParser.
-func NewRawEnvParser() MagicEnvParser {
-	return &rawEnvParser{}
-}
-
-// Match checks if the value matches the raw parser.
-func (*rawEnvParser) Match(_ string, _ string) bool {
-	return true
-}
-
-// Gen generates a raw environment variable.
-func (*rawEnvParser) Gen(k string, v string) *corev1.EnvVar {
-	return &corev1.EnvVar{
-		Name:  k,
-		Value: v,
-	}
-}
-
-// secretEnvParser is a parser for secret-based environment variables.
-type secretEnvParser struct {
-	prefix string
-}
-
-// NewSecretEnvParser creates a new instance of SecretEnvParser.
-func NewSecretEnvParser() MagicEnvParser {
-	return &secretEnvParser{
-		prefix: "secret://",
-	}
-}
-
-// Match checks if the value matches the secret parser.
-func (p *secretEnvParser) Match(_ string, v string) bool {
-	return strings.HasPrefix(v, p.prefix)
-}
-
-// Gen generates a secret-based environment variable.
-func (p *secretEnvParser) Gen(k string, v string) *corev1.EnvVar {
-	vv := strings.TrimPrefix(v, p.prefix)
-	vs := strings.Split(vv, "/")
-	if len(vs) != 2 {
-		return nil
-	}
-
-	return &corev1.EnvVar{
-		Name: k,
-		ValueFrom: &corev1.EnvVarSource{
-			SecretKeyRef: &corev1.SecretKeySelector{
-				LocalObjectReference: corev1.LocalObjectReference{
-					Name: vs[0],
-				},
-				Key: vs[1],
-			},
-		},
-	}
-}
-
-// configMapEnvParser is a parser for configmap-based environment
-// variables.
-type configMapEnvParser struct {
-	prefix string
-}
-
-// NewConfigMapEnvParser creates a new instance of ConfigMapEnvParser.
-func NewConfigMapEnvParser() MagicEnvParser {
-	return &configMapEnvParser{
-		prefix: "configmap://",
-	}
-}
-
-// Match checks if the value matches the configmap parser.
-func (p *configMapEnvParser) Match(_ string, v string) bool {
-	return strings.HasPrefix(v, p.prefix)
-}
-
-// Gen generates a configmap-based environment variable.
-func (p *configMapEnvParser) Gen(k string, v string) *corev1.EnvVar {
-	vv := strings.TrimPrefix(v, p.prefix)
-	vs := strings.Split(vv, "/")
-	if len(vs) != 2 {
-		return nil
-	}
-
-	return &corev1.EnvVar{
-		Name: k,
-		ValueFrom: &corev1.EnvVarSource{
-			ConfigMapKeyRef: &corev1.ConfigMapKeySelector{
-				LocalObjectReference: corev1.LocalObjectReference{
-					Name: vs[0],
-				},
-				Key: vs[1],
-			},
-		},
 	}
 }
